Move files to codeql-cli/distribution directory
