+++ extensions/ql-vscode/src/codeql-cli/distribution/github-api-error.ts
