+++ extensions/ql-vscode/src/codeql-cli/distribution/release.ts
