+++ extensions/ql-vscode/src/codeql-cli/distribution/releases-api-consumer.ts
